New Segurity and fix old problemens

+++ src/main/java/ti/proyectoinia/business/repositories/LoteRepository.java
@@ -6,6 +6,6 @@ import ti.proyectoinia.business.entities.Lote;
 import java.util.List;
 
 @Repository
-public interface LoteRepository extends JpaRepository<Lote, Integer> {
+public interface LoteRepository extends JpaRepository<Lote, Long> {
     List<Lote> findByActivoTrue();
 }
